components/embedder/requirements.txt
einops==0.8.2
fastapi==0.129.0
numpy==2.4.2
pillow==12.1.1
python-dotenv==1.1.0
python-multipart==0.0.22
requests==2.32.5
timm==1.0.26
torch==2.9.0
torchvision==0.24.0
transformers==4.57.6
uvicorn==0.40.0